Replace Scrolling alignment with Tail plugin
