+++ Cargo.toml
@@ -28,6 +28,10 @@ name = "plugin"
 required-features = ["plugin"]
 
 [[example]]
+name = "plugin-tail"
+required-features = ["plugin"]
+
+[[example]]
 name = "styles-plugin"
 required-features = ["plugin"]
