+++ README.md
@@ -15,7 +15,6 @@ This crate provides a configurable `TextBox` to render multiline text inside a b
      - `Top`
      - `Middle`
      - `Bottom`
-     - `Scrolling`
 
 `TextBox` also supports some special characters not handled by embedded-graphics' `Text`:
  - non-breaking space (`\u{200b}`)
