+++ src/alignment/test/mod.rs
@@ -2,5 +2,4 @@ mod center;
 mod justified;
 mod left;
 mod right;
-mod scrolling;
 mod top;
